requirements.txt
langchain
langchain-core
langchain-openai
python-dotenv